FIX - Player models
player models were expecting an Int for scorePerGame, but was receiving a Double. Model changed accordingly

+++ app/src/main/java/com/louie/warzonestats/models/player/PropertiesXXXXXXXXXXXXXXXXXXXXXXXX.kt
@@ -31,7 +31,7 @@ data class PropertiesXXXXXXXXXXXXXXXXXXXXXXXX(
     val objectiveReviver: Int,
     val objectiveTeamWiped: Int,
     val score: Int,
-    val scorePerGame: Int,
+    val scorePerGame: Double,
     val scorePerMinute: Double,
     val timePlayed: Int,
     val wallBangs: Int
